remove `getType()` method.
