+++ metrics/interface/src/main/java/org/apache/iotdb/metrics/metricsets/predefined/jvm/JvmGcMetrics.java
@@ -17,11 +17,10 @@
  * under the License.
  */
 
-package org.apache.iotdb.metrics.predefined.jvm;
+package org.apache.iotdb.metrics.metricsets.predefined.jvm;
 
 import org.apache.iotdb.metrics.AbstractMetricManager;
-import org.apache.iotdb.metrics.predefined.IMetricSet;
-import org.apache.iotdb.metrics.predefined.PredefinedMetric;
+import org.apache.iotdb.metrics.metricsets.IMetricSet;
 import org.apache.iotdb.metrics.type.Counter;
 import org.apache.iotdb.metrics.type.Timer;
 import org.apache.iotdb.metrics.utils.MetricLevel;
@@ -231,11 +230,6 @@ public class JvmGcMetrics implements IMetricSet, AutoCloseable {
   @Override
   public void close() {
     notificationListenerCleanUpRunnables.forEach(Runnable::run);
-  }
-
-  @Override
-  public PredefinedMetric getType() {
-    return PredefinedMetric.JVM;
   }
 
   enum GcGenerationAge {
